383 star rating calcs update

+++ src/crags/entities/route.entity.ts
@@ -67,7 +67,7 @@ export class Route extends BaseEntity {
   @Column()
   defaultGradingSystemId: string;
 
-  @Column({ type: 'float', nullable: true })
+  @Column({ type: 'int', nullable: true })
   @Field({ nullable: true })
   starRating: number;
